implement sort interface for UnlockHashSlice
