+++ types/signatures.go
@@ -7,10 +7,7 @@ package types
 // called 'UnlockConditions'.
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
-	"fmt"
 
 	"github.com/NebulousLabs/Sia/crypto"
 	"github.com/NebulousLabs/Sia/encoding"
@@ -34,13 +31,6 @@ var (
 	ErrSortedUniqueViolation     = errors.New("sorted unique violation")
 	ErrUnlockHashWrongLen        = errors.New("marshalled unlock hash is the wrong length")
 	ErrWholeTransactionViolation = errors.New("covered fields violation")
-
-	// UnlockHashChecksumSize is the size of the checksum used to verify
-	// human-readable addresses. It is not a crypytographically secure
-	// checksum, it's merely intended to prevent typos. 6 is chosen because it
-	// brings the total size of the address to 38 bytes, leaving 2 bytes for
-	// potential version additions in the future.
-	UnlockHashChecksumSize = 6
 
 	// FullCoveredFields is a covered fileds object where the
 	// 'WholeTransaction' field has been set to true. The primary purpose of
@@ -122,12 +112,6 @@ type (
 		PublicKeys         []SiaPublicKey `json:"publickeys"`
 		SignaturesRequired uint64         `json:"signaturesrequired"`
 	}
-
-	// An UnlockHash is a specially constructed hash of the UnlockConditions
-	// type. "Locked" values can be unlocked by providing the UnlockConditions
-	// that hash to a given UnlockHash. See SpendConditions.UnlockHash for
-	// details on how the UnlockHash is constructed.
-	UnlockHash crypto.Hash
 
 	// Each input has a list of public keys and a required number of signatures.
 	// inputSignatures keeps track of which public keys have been used and how many
@@ -402,63 +386,6 @@ func (t *Transaction) validSignatures(currentHeight BlockHeight) error {
 			return ErrMissingSignatures
 		}
 	}
-
-	return nil
-}
-
-// MarshalJSON is implemented on the unlock hash to always produce a hex string
-// upon marshalling.
-func (uh UnlockHash) MarshalJSON() ([]byte, error) {
-	return json.Marshal(uh.String())
-}
-
-// UnmarshalJSON is implemented on the unlock hash to recover an unlock hash
-// that has been encoded to a hex string.
-func (uh *UnlockHash) UnmarshalJSON(b []byte) error {
-	// Check the length of b.
-	if len(b) != crypto.HashSize*2+UnlockHashChecksumSize*2+2 && len(b) != crypto.HashSize*2+2 {
-		return ErrUnlockHashWrongLen
-	}
-	return uh.LoadString(string(b[1 : len(b)-1]))
-}
-
-// String returns the hex representation of the unlock hash as a string - this
-// includes a checksum.
-func (uh UnlockHash) String() string {
-	uhChecksum := crypto.HashObject(uh)
-	return fmt.Sprintf("%x%x", uh[:], uhChecksum[:UnlockHashChecksumSize])
-}
-
-// LoadString loads a hex representation (including checksum) of an unlock hash
-// into an unlock hash object. An error is returned if the string is invalid or
-// fails the checksum.
-func (uh *UnlockHash) LoadString(strUH string) error {
-	// Check the length of strUH.
-	if len(strUH) != crypto.HashSize*2+UnlockHashChecksumSize*2 && len(strUH) != crypto.HashSize*2 {
-		return ErrUnlockHashWrongLen
-	}
-
-	// Decode the unlock hash.
-	var byteUnlockHash []byte
-	var checksum []byte
-	_, err := fmt.Sscanf(strUH[:crypto.HashSize*2], "%x", &byteUnlockHash)
-	if err != nil {
-		return err
-	}
-	// Decode the checksum, if provided.
-	if len(strUH) == crypto.HashSize*2+UnlockHashChecksumSize*2 {
-		_, err = fmt.Sscanf(strUH[crypto.HashSize*2:], "%x", &checksum)
-		if err != nil {
-			return err
-		}
-
-		// Verify the checksum - leave uh as-is unless the checksum is valid.
-		expectedChecksum := crypto.HashBytes(byteUnlockHash)
-		if !bytes.Equal(expectedChecksum[:UnlockHashChecksumSize], checksum) {
-			return ErrInvalidUnlockHashChecksum
-		}
-	}
-	copy(uh[:], byteUnlockHash[:])
 
 	return nil
 }
